requirements.txt
asgiref==3.7.2
chardet==5.2.0
Django==5.0
django-bootstrap-form==3.4
django-factory-boy==1.0.0
django-phonenumber-field==7.2.0
et-xmlfile==1.1.0
factory-boy==3.3.0
Faker==21.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.4
parameterized==0.9.0
phonenumberslite==8.13.27
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyxirr==0.10.1
reportlab==4.0.8
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
whitenoise==6.6.0